+++ frontend/src/pages/homePage/components/mainGraph/MainGraph.style.ts
