+++ frontend/src/pages/homePage/components/mainGraph/MainGraph.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import GraphMenu from '../graphMenu/GraphMenu';
-import S from './MainPageGraph.style';
+import S from './MainGraph.style';
 
 const MainPageGraph = () => {
   const [tabIndex, setTabIndex] = useState(0);
